Finished first draft for FaceByFace model
It got complicated again when it comes to aperture-group and how they can be introduced in a face by face basis approach. Interestingly enough it is how the workflow in Honeybee plugin is designed but it turns out that it doesn't translate to these schemas.

Very close to finish them all. One more push and we will have a first release candidate.