changed parent to be an object
For now I did not make `id` required since faces don't have an id. The sam goes to zones.

In case of writing to file we really only need the name of the parent object but if we want to fetch the parent object we will need more information. Not sure if there will be such a case but there should be. Right?

cc: @antoinedao