feat(api): changed everything! (#17)
* Started documenting model object

* Finished static model

The next round will add dynamic model and more properties to also support an energy model creation.

* session > task

* Added the schema for state

I'm not currently happy with this schema and I think it can be more slim but I'm going to commit the schem as it is so I can move on! I will get back to it soon before sending the PR.

* renamed matrix to tmtx

transmission matrix

* renamed State to ModelState

* restructured schema folders

* Added place holder for main model schemas

* Finished first draft for FaceByFace model

It got complicated again when it comes to aperture-group and how they can be introduced in a face by face basis approach. Interestingly enough it is how the workflow in Honeybee plugin is designed but it turns out that it doesn't translate to these schemas.

Very close to finish them all. One more push and we will have a first release candidate.

* Another round of working on Model objects

It's almost done! Will add examples for each object tomorrow and then add paths.

* Added SpaceBySpaceModel

* Finished Model schemas

* started documenting model paths

* Added MVP paths for model

Next step is to add pagination to get  calls.

* Added radiance materials

This should do it for the MVP products.

* Added type to Radaince materials

* Added two sample JSON files for models

* Changed face_type from integers to string

Addresses thdi comment:

https://github.com/pollination/api/pull/8#discussion_r262762577

* No empty model is allowed!

Addresses this comment:

https://github.com/pollination/api/pull/8#discussion_r262693125

* changed parent to be an object

For now I did not make `id` required since faces don't have an id. The sam goes to zones.

In case of writing to file we really only need the name of the parent object but if we want to fetch the parent object we will need more information. Not sure if there will be such a case but there should be. Right?

cc: @antoinedao

* Added id to all model and face schema

- added id to all Model and Face schema
- removed is_aperture from faces.

* Added base_type and id to materials

* Added Opaque and Transparent to rad_modifiers

* enabled swaggerUI

I don't personally use the UI but it will be useful for the Energy Model schema development.

* vertices > vertex

Resolves https://github.com/pollination/api/issues/13

* firt try on daylight factor simulation object

This is the first try to create the schema for a daylight-factor simulation. Right now I'm happy with how `after_success` property is designed and user can send the postprocess tasks with the original task. Still working on responses.

* commit as is

I'll start from model branch and make all the changes neccessary before I get back to this branch. Will send everything together as a giant PR since we have already discussed what needs to be changed.

* Improved model schema and end points

Added a LinkedModel object to address the need for returning a light-weight model while providing all the needed information.

* Fixed Radiance modifier schemas

* cleaned schemas

Did a lot of cleaning including but not limited to:

- removed simulation end points
- added workflows
- almost finalized the schema for simulation, workflow and outputs

* finished WorkflowStatus

* Added placeholder for zones to FaceByFaceModel

* Added missing model end points

next step is to add end points for simulation

* Added pagination

- added pagination to description
- added link to response headers
- updated sensorgrid end points

* Finished improving model schemas and end points

A very necessary review of each and every schema that I wrote for models. Removed unnecessary end points and added missed ones!

Workflow is the next and last one!

* Finished simulation objects and end points

I changed my mind about `simulation` objects. Enforcing an extra step to create a workflow will be wasteful for simple simulations and adds unneccessary complication on Argo jobs too. Now every simulation is executable and provides end points for controlling the simulation (start/ stop) as well as requesting outputs. I need to replicate these end points for workflows. Workflows will reference back to simulations for outputs.

The main advantage of workflows is passing the values between the simulations.  It can also be used to execute several simulations at the same time. Hello design explorer - we should probably call it pollinator! :sunflower: :bee:

* Updated workflows based on new simulation schema

This should be the last commit before merging it back to pollination/api.

I'm still not quite sure about the current approach to outputs access but what we have is way over than what can be called an MVP product! So let merge this and we will iron out the rest of the details and we are spinning up the services.

:rocket: :v:

* Fixed mistakes in schema

* Added pagination fields to query parameters