Added LICENSE to .gitignore
This branch will be forked out to ladybug-tools-in2 and we will need to change the license to BSD on the other repository. Adding the LICENSE to gitignore will avoid future merge conflicts. Since MIT and BSD are two way compatible this will also work fine from licensing point of view.